Move value object to new location so we can build out more of them
