+++ service-api/app/test/AppTest/Service/ViewerCodes/ViewerCodeServiceTest.php
@@ -8,8 +8,8 @@ use App\DataAccess\Repository\KeyCollisionException;
 use App\DataAccess\Repository\UserLpaActorMapInterface;
 use App\DataAccess\Repository\ViewerCodeActivityInterface;
 use App\DataAccess\Repository\ViewerCodesInterface;
-use App\Entity\Value\LpaUid;
 use App\Service\ViewerCodes\ViewerCodeService;
+use App\Value\LpaUid;
 use DateTime;
 use DateTimeImmutable;
 use DateTimeZone;
