+++ service-api/app/test/AppTest/Value/LpaUidTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AppTest\Entity\Value;
+namespace AppTest\Value;
 
-use App\Entity\Value\LpaUid;
 use App\Enum\LpaSource;
+use App\Value\LpaUid;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
